Testing in PHP7
Now that we have support for PHP7 in SS 3.x, we should test the Postgres module against it.
I've set up the tests to run weekly on this branch (and master) via TravisCI new cron job feature.
While we have a nightly "all modules" run in silverstripe-installer, it doesn't have the same matrix coverage as this run here.

I've also removed removed tests against older (and unsupported) 3.x releases (it was testing against 3.2)